fix: 검색 시 키보드가 내려가지 않도록 수정
