+++ feature-ui-search/src/main/kotlin/team/duckie/app/android/feature/ui/search/viewmodel/sideeffect/SearchSideEffect.kt
@@ -15,7 +15,4 @@ internal sealed class SearchSideEffect {
      * @param exception 발생한 예외
      */
     class ReportError(val exception: Throwable) : SearchSideEffect()
-
-    /** 검색 화면에서 키보드를 내립니다. */
-    object HideKeyBoard : SearchSideEffect()
 }
